requirements.txt
python-telegram-bot>=20.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
python-telegram-bot-calendar>=1.0.5
